ROFO-200 refactor: MapScreen 동작을 HomeScreen으로 이전
