+++ presentation/src/main/java/com/weit2nd/presentation/ui/home/HomeIntent.kt
@@ -1,6 +1,26 @@
 package com.weit2nd.presentation.ui.home
 
+import com.kakao.vectormap.KakaoMap
+import com.kakao.vectormap.LatLng
+
 sealed class HomeIntent {
+    data class RequestFoodSpots(
+        val centerLat: Double,
+        val centerLng: Double,
+    ) : HomeIntent()
+
+    data class RefreshMarkers(
+        val map: KakaoMap,
+    ) : HomeIntent()
+
+    data class RequestCameraMove(
+        val position: LatLng,
+    ) : HomeIntent()
+
+    data class ShowFoodSpotSummary(
+        val foodSpotId: Long,
+    ) : HomeIntent()
+
     data object NavToFoodSpotReport : HomeIntent()
 
     data object NavToBack : HomeIntent()
@@ -12,4 +32,6 @@ sealed class HomeIntent {
     ) : HomeIntent()
 
     data object NavToMyPage : HomeIntent()
+
+    data object ShowRetryButton : HomeIntent()
 }
