+++ presentation/src/main/java/com/weit2nd/presentation/ui/home/HomeScreen.kt
@@ -1,13 +1,16 @@
 package com.weit2nd.presentation.ui.home
 
+import android.util.Log
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
@@ -20,23 +23,39 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberUpdatedState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.viewinterop.AndroidView
 import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleEventObserver
+import androidx.lifecycle.LifecycleOwner
 import coil.compose.AsyncImage
+import com.kakao.vectormap.KakaoMap
+import com.kakao.vectormap.KakaoMapReadyCallback
+import com.kakao.vectormap.LatLng
+import com.kakao.vectormap.MapLifeCycleCallback
+import com.kakao.vectormap.MapView
+import com.kakao.vectormap.camera.CameraUpdateFactory
+import com.kakao.vectormap.label.LabelOptions
+import com.kakao.vectormap.label.LabelStyle
+import com.kakao.vectormap.label.LabelStyles
 import com.weit2nd.presentation.R
 import com.weit2nd.presentation.navigation.dto.PlaceSearchDTO
-import com.weit2nd.presentation.ui.common.SearchTopBar
-import com.weit2nd.presentation.ui.map.MapScreen
+import com.weit2nd.presentation.ui.common.currentposition.CurrentPositionBtn
 import com.weit2nd.presentation.ui.theme.Gray1
 import com.weit2nd.presentation.ui.theme.Gray2
 import com.weit2nd.presentation.ui.theme.RoadyFoodyTheme
@@ -70,31 +89,52 @@ fun HomeScreen(
             HomeSideEffect.NavToMyPage -> {
                 navToMyPage()
             }
+            is HomeSideEffect.RefreshMarkers -> {
+                drawMarkers(sideEffect.map, sideEffect.foodSpotMarkers)
+            }
+            is HomeSideEffect.MoveCamera -> {
+                moveCamera(sideEffect.map, sideEffect.position)
+            }
         }
     }
-    Scaffold(
-        topBar = {
-            SearchTopBar(
-                modifier = Modifier.background(Color.White),
-                searchWords = state.searchWords,
-                textFieldEnabled = false,
-                readOnly = true,
-                onTextFieldClick = vm::onSearchPlaceClick,
-                onNavigationButtonClick = vm::onNavigateButtonClick,
-            )
-        },
-    ) {
+
+    val context = LocalContext.current
+    val mapView =
+        remember {
+            MapView(context).apply {
+                start(
+                    mapLifeCycleCallback(),
+                    kakaoMapReadyCallback(
+                        onMapReady = vm::onMapReady,
+                        onCameraMoveEnd = vm::onCameraMoveEnd,
+                        onMarkerClick = vm::onMarkerClick,
+                        position = state.initialLatLng,
+                    ),
+                )
+            }
+        }
+
+    LaunchedEffect(state.foodSpots) {
+        state.map?.let {
+            drawMarkers(it, state.foodSpots)
+        }
+    }
+
+    DisposableEffectWithLifeCycle(
+        onResume = mapView::resume,
+        onPause = mapView::pause,
+    )
+
+    Scaffold {
         Surface(
             modifier =
                 Modifier
                     .fillMaxSize()
                     .padding(it),
         ) {
-            MapScreen(
+            AndroidView(
                 modifier = Modifier.fillMaxSize(),
-                initialPosition = state.initialLatLng,
-                onCameraMoveEnd = vm::onCameraMoved,
-                onMarkerClick = vm::onFoodSpotMarkerClick,
+                factory = { mapView },
             )
             Box(
                 modifier =
@@ -102,20 +142,44 @@ fun HomeScreen(
                         .fillMaxSize()
                         .padding(16.dp),
             ) {
-                SearchBar(
+                Column(
                     modifier =
                         Modifier
                             .align(Alignment.TopCenter)
-                            .fillMaxWidth()
-                            .clip(RoundedCornerShape(20.dp))
-                            .background(MaterialTheme.colorScheme.surface),
-                    searchWords = state.searchWords,
-                    profileImage = state.profileImage,
-                    onSearchBarClick = vm::onSearchPlaceClick,
-                    onProfileClick = vm::onProfileClick,
+                            .fillMaxWidth(),
+                    horizontalAlignment = Alignment.CenterHorizontally,
+                ) {
+                    SearchBar(
+                        modifier =
+                            Modifier
+                                .fillMaxWidth()
+                                .clip(RoundedCornerShape(20.dp))
+                                .background(MaterialTheme.colorScheme.surface),
+                        searchWords = state.searchWords,
+                        profileImage = state.profileImage,
+                        onSearchBarClick = vm::onSearchPlaceClick,
+                        onProfileClick = vm::onProfileClick,
+                    )
+                    if (state.isMoved) {
+                        Spacer(modifier = Modifier.height(16.dp))
+                        Button(
+                            onClick = {
+                                state.map?.let { map -> vm.onClickRefreshFoodSpotBtn(map) }
+                            },
+                        ) {
+                            Text(text = "이 위치에서 검색")
+                        }
+                    }
+                }
+                CurrentPositionBtn(
+                    modifier =
+                        Modifier
+                            .size(40.dp)
+                            .align(Alignment.BottomStart),
+                    onClick = vm::onClickCurrentPositionBtn,
                 )
                 FoodSpotReportButton(
-                    modifier = Modifier.align(Alignment.BottomStart),
+                    modifier = Modifier.align(Alignment.BottomEnd),
                     onClick = vm::onClickReportBtn,
                 )
             }
@@ -217,6 +281,96 @@ private fun SearchBar(
             )
         }
     }
+}
+
+private fun drawMarkers(
+    map: KakaoMap,
+    foodSpots: List<FoodSpotState>,
+    isRefresh: Boolean = true,
+) {
+    if (isRefresh) {
+        map.labelManager?.layer?.removeAll()
+    }
+    foodSpots.forEach {
+        val styles =
+            map.labelManager
+                ?.addLabelStyles(LabelStyles.from(LabelStyle.from(android.R.drawable.star_on)))
+        val options =
+            LabelOptions
+                .from(it.position)
+                .setStyles(styles)
+                .apply {
+                    labelId = it.id.toString()
+                }
+        options.labelId = it.id.toString()
+        map.labelManager?.layer?.addLabel(options)
+    }
+}
+
+private fun mapLifeCycleCallback() =
+    object : MapLifeCycleCallback() {
+        override fun onMapDestroy() {
+            Log.d("KakaoMap", "onMapDestroy")
+        }
+
+        override fun onMapError(error: Exception) {
+            Log.e("KakaoMap", "onMapError: ", error)
+        }
+    }
+
+@Composable
+private fun DisposableEffectWithLifeCycle(
+    // 오류로 compose.ui의 LocalLifecycleOwner 사용
+    lifecycleOwner: LifecycleOwner = androidx.compose.ui.platform.LocalLifecycleOwner.current,
+    onResume: () -> Unit,
+    onPause: () -> Unit,
+) {
+    val currentOnResume by rememberUpdatedState(onResume)
+    val currentOnPause by rememberUpdatedState(onPause)
+
+    DisposableEffect(lifecycleOwner) {
+        val observer =
+            LifecycleEventObserver { _, event ->
+                if (event == Lifecycle.Event.ON_RESUME) {
+                    currentOnResume()
+                } else if (event == Lifecycle.Event.ON_PAUSE) {
+                    currentOnPause()
+                }
+            }
+
+        lifecycleOwner.lifecycle.addObserver(observer)
+
+        onDispose {
+            lifecycleOwner.lifecycle.removeObserver(observer)
+        }
+    }
+}
+
+private fun kakaoMapReadyCallback(
+    onMapReady: (KakaoMap) -> Unit,
+    onCameraMoveEnd: (LatLng) -> Unit,
+    onMarkerClick: (Long) -> Unit,
+    position: LatLng,
+) = object : KakaoMapReadyCallback() {
+    override fun onMapReady(map: KakaoMap) {
+        onMapReady(map)
+        map.setOnCameraMoveEndListener { _, currentPosition, _ ->
+            onCameraMoveEnd(currentPosition.position)
+        }
+        map.setOnLabelClickListener { _, _, label ->
+            onMarkerClick(label.labelId.toLong())
+        }
+    }
+
+    override fun getPosition(): LatLng = position
+}
+
+private fun moveCamera(
+    map: KakaoMap,
+    position: LatLng,
+) {
+    val cameraUpdate = CameraUpdateFactory.newCenterPosition(position)
+    map.moveCamera(cameraUpdate)
 }
 
 @Preview
