+++ presentation/src/main/java/com/weit2nd/presentation/ui/home/HomeViewModel.kt
@@ -1,18 +1,24 @@
 package com.weit2nd.presentation.ui.home
 
+import android.util.Log
 import androidx.lifecycle.SavedStateHandle
+import com.kakao.vectormap.KakaoMap
 import com.kakao.vectormap.LatLng
+import com.weit2nd.domain.model.Coordinate
+import com.weit2nd.domain.usecase.search.SearchFoodSpotsUseCase
 import com.weit2nd.presentation.base.BaseViewModel
 import com.weit2nd.presentation.navigation.HomeNavRoutes
 import com.weit2nd.presentation.navigation.dto.CoordinateDTO
 import com.weit2nd.presentation.navigation.dto.PlaceSearchDTO
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.Job
 import org.orbitmvi.orbit.viewmodel.container
 import javax.inject.Inject
 
 @HiltViewModel
 class HomeViewModel @Inject constructor(
     savedStateHandle: SavedStateHandle,
+    private val searchFoodSpotsUseCase: SearchFoodSpotsUseCase,
 ) : BaseViewModel<HomeState, HomeSideEffect>() {
     private val placeSearch = savedStateHandle.get<PlaceSearchDTO>(HomeNavRoutes.PLACE_SEARCH_KEY)
 
@@ -27,6 +33,7 @@ class HomeViewModel @Inject constructor(
             ),
         )
     private var currentCameraPosition: LatLng = container.stateFlow.value.initialLatLng
+    private var searchFoodSpotsJob: Job = Job().apply { complete() }
 
     fun onClickReportBtn() {
         HomeIntent.NavToFoodSpotReport.post()
@@ -40,16 +47,41 @@ class HomeViewModel @Inject constructor(
         HomeIntent.NavToSearch.post()
     }
 
-    fun onCameraMoved(position: LatLng) {
-        currentCameraPosition = position
-    }
-
-    fun onFoodSpotMarkerClick(foodSpotId: Long) {
-        HomeIntent.NavToFoodSpotDetail(foodSpotId).post()
-    }
-
     fun onProfileClick() {
         HomeIntent.NavToMyPage.post()
+    }
+
+    fun onMapReady(kakaoMap: KakaoMap) {
+        HomeIntent.RefreshMarkers(kakaoMap).post()
+    }
+
+    fun onClickCurrentPositionBtn(currentPosition: LatLng) {
+        HomeIntent.RequestCameraMove(currentPosition).post()
+    }
+
+    fun onMarkerClick(foodSpotId: Long) {
+        HomeIntent.ShowFoodSpotSummary(foodSpotId).post()
+    }
+
+    fun onCameraMoveEnd(currentPosition: LatLng) {
+        currentCameraPosition = currentPosition
+        HomeIntent.ShowRetryButton.post()
+    }
+
+    fun onClickRefreshFoodSpotBtn(map: KakaoMap) {
+        val viewport = map.viewport
+        val x = viewport.width() / 2
+        val y = viewport.height() / 2
+        val centerPosition = map.fromScreenPoint(x, y)
+        if (centerPosition != null) {
+            searchFoodSpotsJob.cancel()
+            searchFoodSpotsJob =
+                HomeIntent
+                    .RequestFoodSpots(
+                        centerLat = centerPosition.latitude,
+                        centerLng = centerPosition.longitude,
+                    ).post()
+        }
     }
 
     private fun HomeIntent.post() =
@@ -81,6 +113,65 @@ class HomeViewModel @Inject constructor(
                 }
                 HomeIntent.NavToMyPage -> {
                     postSideEffect(HomeSideEffect.NavToMyPage)
+                }
+
+                is HomeIntent.RefreshMarkers -> {
+                    reduce {
+                        state.copy(
+                            map = map,
+                        )
+                    }
+                    postSideEffect(
+                        HomeSideEffect.RefreshMarkers(
+                            map,
+                            state.foodSpots,
+                        ),
+                    )
+                }
+                is HomeIntent.RequestCameraMove -> {
+                    state.map?.let { map ->
+                        postSideEffect(HomeSideEffect.MoveCamera(map, position))
+                    }
+                }
+                is HomeIntent.RequestFoodSpots -> {
+                    reduce {
+                        state.copy(
+                            isMoved = false,
+                        )
+                    }
+                    runCatching {
+                        // TODO 이름, 카테고리 가져오기
+                        // TODO radius를 넣을 때 유저 레벨?을 계산해서 넣기
+                        val foodSpots =
+                            searchFoodSpotsUseCase
+                                .invoke(
+                                    centerCoordinate =
+                                        Coordinate(
+                                            longitude = centerLng,
+                                            latitude = centerLat,
+                                        ),
+                                    radius = 500,
+                                    name = null,
+                                    categoryIds = emptyList(),
+                                ).map {
+                                    it.toFoodSpotState()
+                                }
+                        reduce {
+                            state.copy(
+                                foodSpots = foodSpots,
+                            )
+                        }
+                    }.onFailure { exception ->
+                        Log.e("RequestFoodSpotsFail", "${exception.message}")
+                    }
+                }
+                is HomeIntent.ShowFoodSpotSummary -> TODO()
+                HomeIntent.ShowRetryButton -> {
+                    reduce {
+                        state.copy(
+                            isMoved = true,
+                        )
+                    }
                 }
             }
         }
