--- presentation/src/main/java/com/weit2nd/presentation/ui/map/MapScreen.kt
@@ -1,203 +1,0 @@
-package com.weit2nd.presentation.ui.map
-
-import android.util.Log
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Button
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberUpdatedState
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.viewinterop.AndroidView
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleEventObserver
-import androidx.lifecycle.LifecycleOwner
-import com.kakao.vectormap.KakaoMap
-import com.kakao.vectormap.KakaoMapReadyCallback
-import com.kakao.vectormap.LatLng
-import com.kakao.vectormap.MapLifeCycleCallback
-import com.kakao.vectormap.MapView
-import com.kakao.vectormap.camera.CameraUpdateFactory
-import com.kakao.vectormap.label.LabelOptions
-import com.kakao.vectormap.label.LabelStyle
-import com.kakao.vectormap.label.LabelStyles
-import com.weit2nd.presentation.ui.common.currentposition.CurrentPositionBtn
-import org.orbitmvi.orbit.compose.collectAsState
-import org.orbitmvi.orbit.compose.collectSideEffect
-
-@Composable
-fun MapScreen(
-    modifier: Modifier = Modifier,
-    vm: MapViewModel = hiltViewModel(),
-    initialPosition: LatLng = LatLng.from(37.5597706, 126.9423666),
-    onCameraMoveEnd: (LatLng) -> Unit,
-    onMarkerClick: (foodSpotId: Long) -> Unit,
-) {
-    val state = vm.collectAsState()
-    vm.collectSideEffect { sideEffect ->
-        when (sideEffect) {
-            is MapSideEffect.RefreshMarkers -> {
-                drawMarkers(sideEffect.map, sideEffect.foodSpotMarkers)
-            }
-
-            is MapSideEffect.MoveCamera -> {
-                moveCamera(sideEffect.map, sideEffect.position)
-            }
-
-            is MapSideEffect.SendCameraPosition -> {
-                onCameraMoveEnd(sideEffect.position)
-            }
-
-            is MapSideEffect.SendClickedFoodSpotId -> {
-                onMarkerClick(sideEffect.foodSpotId)
-            }
-        }
-    }
-    val context = LocalContext.current
-    val mapView =
-        remember {
-            MapView(context).apply {
-                start(
-                    mapLifeCycleCallback(),
-                    kakaoMapReadyCallback(
-                        onMapReady = vm::onMapReady,
-                        onCameraMoveEnd = vm::onCameraMoveEnd,
-                        onMarkerClick = vm::onMarkerClick,
-                        position = initialPosition,
-                    ),
-                )
-            }
-        }
-
-    LaunchedEffect(state.value.foodSpots) {
-        state.value.map?.let {
-            drawMarkers(it, state.value.foodSpots)
-        }
-    }
-
-    DisposableEffectWithLifeCycle(onResume = mapView::resume, onPause = mapView::pause)
-
-    Box(modifier = modifier) {
-        AndroidView(
-            modifier = modifier,
-            factory = { mapView },
-        )
-
-        CurrentPositionBtn(
-            modifier =
-                Modifier
-                    .align(Alignment.BottomEnd)
-                    .padding(end = 16.dp, bottom = 24.dp),
-            onClick = vm::onClickCurrentPositionBtn,
-        )
-
-        if (state.value.isMoved) {
-            Button(
-                modifier = Modifier.align(Alignment.TopCenter).padding(8.dp),
-                onClick = {
-                    state.value.map?.let { vm.onClickRefreshFoodSpotBtn(it) }
-                },
-            ) {
-                Text(text = "이 위치에서 검색")
-            }
-        }
-    }
-}
-
-private fun moveCamera(
-    map: KakaoMap,
-    position: LatLng,
-) {
-    val cameraUpdate = CameraUpdateFactory.newCenterPosition(position)
-    map.moveCamera(cameraUpdate)
-}
-
-private fun mapLifeCycleCallback() =
-    object : MapLifeCycleCallback() {
-        override fun onMapDestroy() {
-            Log.d("KakaoMap", "onMapDestroy")
-        }
-
-        override fun onMapError(error: Exception) {
-            Log.e("KakaoMap", "onMapError: ", error)
-        }
-    }
-
-private fun kakaoMapReadyCallback(
-    onMapReady: (KakaoMap) -> Unit,
-    onCameraMoveEnd: (LatLng) -> Unit,
-    onMarkerClick: (Long) -> Unit,
-    position: LatLng,
-) = object : KakaoMapReadyCallback() {
-    override fun onMapReady(map: KakaoMap) {
-        onMapReady(map)
-        map.setOnCameraMoveEndListener { _, cameraPosition, _ ->
-            onCameraMoveEnd(cameraPosition.position)
-        }
-        map.setOnLabelClickListener { _, _, label ->
-            onMarkerClick(label.labelId.toLong())
-        }
-    }
-
-    override fun getPosition(): LatLng = position
-}
-
-private fun drawMarkers(
-    map: KakaoMap,
-    foodSpots: List<FoodSpotState>,
-    isRefresh: Boolean = true,
-) {
-    if (isRefresh) {
-        map.labelManager?.layer?.removeAll()
-    }
-    foodSpots.forEach {
-        val styles =
-            map.labelManager
-                ?.addLabelStyles(LabelStyles.from(LabelStyle.from(android.R.drawable.star_on)))
-        val options =
-            LabelOptions
-                .from(it.position)
-                .setStyles(styles)
-                .apply {
-                    labelId = it.id.toString()
-                }
-        options.labelId = it.id.toString()
-        map.labelManager?.layer?.addLabel(options)
-    }
-}
-
-@Composable
-private fun DisposableEffectWithLifeCycle(
-    // 오류로 compose.ui의 LocalLifecycleOwner 사용
-    lifecycleOwner: LifecycleOwner = androidx.compose.ui.platform.LocalLifecycleOwner.current,
-    onResume: () -> Unit,
-    onPause: () -> Unit,
-) {
-    val currentOnResume by rememberUpdatedState(onResume)
-    val currentOnPause by rememberUpdatedState(onPause)
-
-    DisposableEffect(lifecycleOwner) {
-        val observer =
-            LifecycleEventObserver { _, event ->
-                if (event == Lifecycle.Event.ON_RESUME) {
-                    currentOnResume()
-                } else if (event == Lifecycle.Event.ON_PAUSE) {
-                    currentOnPause()
-                }
-            }
-
-        lifecycleOwner.lifecycle.addObserver(observer)
-
-        onDispose {
-            lifecycleOwner.lifecycle.removeObserver(observer)
-        }
-    }
-}
