--- presentation/src/main/java/com/weit2nd/presentation/ui/map/MapState.kt
@@ -1,9 +1,0 @@
-package com.weit2nd.presentation.ui.map
-
-import com.kakao.vectormap.KakaoMap
-
-data class MapState(
-    val map: KakaoMap? = null,
-    val foodSpots: List<FoodSpotState> = emptyList(),
-    val isMoved: Boolean = false,
-)
